Fixed support of the special symbols in path: ' &!~`@#$^(+)_=%-;. Related Issue #88
Back to the old label-goto logic instead of `for` (d4b46cd198) because of more hard escaping for all special symbols above. This is the moment when you finally start to feel sick from batch scripts -_-

This changes also provides new `__p_call` flag - To use the call-type logic when invoking DllExport.

And also adds handling of unknown action. Just not to scare users...

Other details in #88